+++ packages/dependency-extraction-webpack-plugin/lib/util.js
@@ -10,6 +10,7 @@ const BUNDLED_PACKAGES = [
 	'@wordpress/upload-media',
 	'@wordpress/fields',
 	'@wordpress/views',
+	'@wordpress/ui',
 ];
 
 /**