+++ packages/notices/src/store/constants.ts
@@ -1,15 +1,16 @@
 /**
+ * Internal dependencies
+ */
+import type { Notice } from './types';
+
+/**
  * Default context to use for notice grouping when not otherwise specified. Its
  * specific value doesn't hold much meaning, but it must be reasonably unique
  * and, more importantly, referenced consistently in the store implementation.
- *
- * @type {string}
  */
 export const DEFAULT_CONTEXT = 'global';
 
 /**
  * Default notice status.
- *
- * @type {string}
  */
-export const DEFAULT_STATUS = 'info';
+export const DEFAULT_STATUS: Notice[ 'status' ] = 'info';