+++ packages/shortcode/src/index.ts
@@ -3,18 +3,34 @@
  */
 import memize from 'memize';
 
+/**
+ * Internal dependencies
+ */
+import type {
+	ShortcodeAttrs,
+	ShortcodeMatch,
+	ShortcodeOptions,
+	Match,
+	ReplaceCallback,
+	ShortcodeInstance,
+} from './types';
+
 export * from './types';
 
 /**
  * Find the next matching shortcode.
  *
- * @param {string} tag   Shortcode tag.
- * @param {string} text  Text to search.
- * @param {number} index Index to start search from.
+ * @param tag   Shortcode tag.
+ * @param text  Text to search.
+ * @param index Index to start search from.
  *
- * @return {import('./types').ShortcodeMatch | undefined} Matched information.
+ * @return Matched information.
  */
-export function next( tag, text, index = 0 ) {
+export function next(
+	tag: string,
+	text: string,
+	index: number = 0
+): ShortcodeMatch | undefined {
 	const re = regexp( tag );
 
 	re.lastIndex = index;
@@ -30,7 +46,7 @@ export function next( tag, text, index = 0 ) {
 		return next( tag, text, re.lastIndex );
 	}
 
-	const result = {
+	const result: ShortcodeMatch = {
 		index: match.index,
 		content: match[ 0 ],
 		shortcode: fromMatch( match ),
@@ -54,17 +70,26 @@ export function next( tag, text, index = 0 ) {
 /**
  * Replace matching shortcodes in a block of text.
  *
- * @param {string}                            tag      Shortcode tag.
- * @param {string}                            text     Text to search.
- * @param {import('./types').ReplaceCallback} callback Function to process the match and return
- *                                                     replacement string.
+ * @param tag      Shortcode tag.
+ * @param text     Text to search.
+ * @param callback Function to process the match and return
+ *                 replacement string.
  *
- * @return {string} Text with shortcodes replaced.
+ * @return Text with shortcodes replaced.
  */
-export function replace( tag, text, callback ) {
+export function replace(
+	tag: string,
+	text: string,
+	callback: ReplaceCallback
+) {
 	return text.replace(
 		regexp( tag ),
-		function ( match, left, $3, attrs, slash, content, closing, right ) {
+		// Let us use spread syntax to capture the arguments object.
+		( ...args: string[] ): string => {
+			const match = args[ 0 ];
+			const left = args[ 1 ];
+			const right = args[ 7 ];
+
 			// If both extra brackets exist, the shortcode has been properly
 			// escaped.
 			if ( left === '[' && right === ']' ) {
@@ -72,7 +97,7 @@ export function replace( tag, text, callback ) {
 			}
 
 			// Create the match object and pass it through the callback.
-			const result = callback( fromMatch( arguments ) );
+			const result = callback( fromMatch( args ) );
 
 			// Make sure to return any of the extra brackets if they weren't used to
 			// escape the shortcode.
@@ -90,12 +115,12 @@ export function replace( tag, text, callback ) {
  * `tag` string, a string or object of `attrs`, a boolean indicating whether to
  * format the shortcode using a `single` tag, and a `content` string.
  *
- * @param {Object} options
+ * @param options Shortcode options.
  *
- * @return {string} String representation of the shortcode.
+ * @return String representation of the shortcode.
  */
-export function string( options ) {
-	return new shortcode( options ).string();
+export function string( options: ShortcodeOptions ): string {
+	return new Shortcode( options ).string();
 }
 
 /**
@@ -114,11 +139,11 @@ export function string( options ) {
  * 6. The closing tag.
  * 7. An extra `]` to allow for escaping shortcodes with double `[[]]`
  *
- * @param {string} tag Shortcode tag.
+ * @param tag Shortcode tag.
  *
- * @return {RegExp} Shortcode RegExp.
+ * @return Shortcode RegExp.
  */
-export function regexp( tag ) {
+export function regexp( tag: string ): RegExp {
 	return new RegExp(
 		'\\[(\\[?)(' +
 			tag +
@@ -142,11 +167,11 @@ export function regexp( tag ) {
  *
  * @param {string} text Serialised shortcode attributes.
  *
- * @return {import('./types').ShortcodeAttrs} Parsed shortcode attributes.
+ * @return {ShortcodeAttrs} Parsed shortcode attributes.
  */
-export const attrs = memize( ( text ) => {
-	const named = {};
-	const numeric = [];
+export const attrs = memize( ( text: string ): ShortcodeAttrs => {
+	const named: Record< string, string | undefined > = {};
+	const numeric: string[] = [];
 
 	// This regular expression is reused from `shortcode_parse_atts()` in
 	// `wp-includes/shortcodes.php`.
@@ -197,12 +222,12 @@ export const attrs = memize( ( text ) => {
  * by `regexp()`. `match` can also be set to the `arguments` from a callback
  * passed to `regexp.replace()`.
  *
- * @param {import('./types').Match} match Match array.
+ * @param match Match array.
  *
- * @return {InstanceType<import('./types').shortcode>} Shortcode instance.
+ * @return Shortcode instance.
  */
-export function fromMatch( match ) {
-	let type;
+export function fromMatch( match: Match ): ShortcodeInstance {
+	let type: 'self-closing' | 'closed' | 'single';
 
 	if ( match[ 4 ] ) {
 		type = 'self-closing';
@@ -212,7 +237,7 @@ export function fromMatch( match ) {
 		type = 'single';
 	}
 
-	return new shortcode( {
+	return new Shortcode( {
 		tag: match[ 2 ],
 		attrs: match[ 3 ],
 		type,
@@ -227,13 +252,27 @@ export function fromMatch( match ) {
  * containing a `tag` string, a string or object of `attrs`, a string indicating
  * the `type` of the shortcode ('single', 'self-closing', or 'closed'), and a
  * `content` string.
- *
- * @type {import('./types').shortcode} Shortcode instance.
  */
-const shortcode = Object.assign(
-	function ( options ) {
-		const { tag, attrs: attributes, type, content } = options || {};
-		Object.assign( this, { tag, type, content } );
+class Shortcode implements ShortcodeInstance {
+	// Instance properties
+	tag: string;
+	type?: 'self-closing' | 'closed' | 'single';
+	content?: string;
+	attrs: ShortcodeAttrs;
+
+	// Static methods
+	static next = next;
+	static replace = replace;
+	static string = string;
+	static regexp = regexp;
+	static attrs = attrs;
+	static fromMatch = fromMatch;
+
+	constructor( options: ShortcodeOptions ) {
+		const { tag, attrs: attributes, type, content } = options;
+		this.tag = tag;
+		this.type = type;
+		this.content = content;
 
 		// Ensure we have a correctly formatted `attrs` object.
 		this.attrs = {
@@ -245,50 +284,46 @@ const shortcode = Object.assign(
 			return;
 		}
 
-		const attributeTypes = [ 'named', 'numeric' ];
-
 		// Parse a string of attributes.
 		if ( typeof attributes === 'string' ) {
 			this.attrs = attrs( attributes );
 			// Identify a correctly formatted `attrs` object.
 		} else if (
-			attributes.length === attributeTypes.length &&
-			attributeTypes.every( ( t, key ) => t === attributes[ key ] )
+			'named' in attributes &&
+			'numeric' in attributes &&
+			attributes.named !== undefined &&
+			attributes.numeric !== undefined
 		) {
-			this.attrs = attributes;
-			// Handle a flat object of attributes.
+			this.attrs = attributes as ShortcodeAttrs;
+			// Handle a flat object of attributes (e.g., { foo: 'bar', baz: 'qux' }).
 		} else {
 			Object.entries( attributes ).forEach( ( [ key, value ] ) => {
-				this.set( key, value );
+				if (
+					typeof value === 'string' ||
+					typeof value === 'undefined'
+				) {
+					this.set( key, value );
+				}
 			} );
 		}
-	},
-	{
-		next,
-		replace,
-		string,
-		regexp,
-		attrs,
-		fromMatch,
 	}
-);
 
-Object.assign( shortcode.prototype, {
 	/**
 	 * Get a shortcode attribute.
 	 *
 	 * Automatically detects whether `attr` is named or numeric and routes it
 	 * accordingly.
 	 *
-	 * @param {(number|string)} attr Attribute key.
+	 * @param attr Attribute key.
 	 *
-	 * @return {string} Attribute value.
+	 * @return Attribute value.
 	 */
-	get( attr ) {
-		return this.attrs[ typeof attr === 'number' ? 'numeric' : 'named' ][
-			attr
-		];
-	},
+	get( attr: string | number ): string | undefined {
+		if ( typeof attr === 'number' ) {
+			return this.attrs.numeric[ attr ];
+		}
+		return this.attrs.named[ attr ];
+	}
 
 	/**
 	 * Set a shortcode attribute.
@@ -296,23 +331,26 @@ Object.assign( shortcode.prototype, {
 	 * Automatically detects whether `attr` is named or numeric and routes it
 	 * accordingly.
 	 *
-	 * @param {(number|string)} attr  Attribute key.
-	 * @param {string}          value Attribute value.
+	 * @param attr  Attribute key.
+	 * @param value Attribute value.
 	 *
-	 * @return {InstanceType< import('./types').shortcode >} Shortcode instance.
+	 * @return Shortcode instance.
 	 */
-	set( attr, value ) {
-		this.attrs[ typeof attr === 'number' ? 'numeric' : 'named' ][ attr ] =
-			value;
+	set( attr: string | number, value: string ): this {
+		if ( typeof attr === 'number' ) {
+			this.attrs.numeric[ attr ] = value;
+		} else {
+			this.attrs.named[ attr ] = value;
+		}
 		return this;
-	},
+	}
 
 	/**
 	 * Transform the shortcode into a string.
 	 *
-	 * @return {string} String representation of the shortcode.
+	 * @return String representation of the shortcode.
 	 */
-	string() {
+	string(): string {
 		let text = '[' + this.tag;
 
 		this.attrs.numeric.forEach( ( value ) => {
@@ -344,7 +382,7 @@ Object.assign( shortcode.prototype, {
 
 		// Add the closing tag.
 		return text + '[/' + this.tag + ']';
-	},
-} );
+	}
+}
 
-export default shortcode;
+export default Shortcode;