+++ packages/ui/src/button/test/button.test.tsx
@@ -74,14 +74,18 @@ describe( 'Button', () => {
 		expect( button ).not.toHaveAttribute( 'aria-disabled', 'true' );
 	} );
 
-	it( 'supports custom render prop while retaining the default accessible when disabled behavior', () => {
+	it( 'supports custom render prop while retaining the default focusable when disabled behavior', () => {
 		render(
+			// Disabling because this lint rule was meant for the
+			// `@wordpress/components` Button, but is being applied here.
+			// TODO: rework the lint rule so that it checks the package
+			// where the Button comes from.
 			// eslint-disable-next-line jsx-a11y/anchor-has-content, no-restricted-syntax
-			<Button render={ <a href="/" /> } disabled>
+			<Button render={ <a href="/" /> } nativeButton={ false } disabled>
 				Click me
 			</Button>
 		);
-		const button = screen.getByRole( 'link', { name: 'Click me' } );
+		const button = screen.getByRole( 'button', { name: 'Click me' } );
 
 		expect( button ).toHaveAttribute( 'aria-disabled', 'true' );
 	} );
@@ -93,7 +97,6 @@ describe( 'Button', () => {
 				Click me
 			</Button>
 		);
-
 		expect(
 			screen.getByRole( 'button', { name: 'Click me' } )
 		).toHaveClass( customClass );